feat(queue-processor): split queue processing out of queue adapter
