+++ test/unit/jobs/adapter/BullQueueAdapter.test.ts
@@ -1,16 +1,9 @@
 import Bull from 'bull';
 import { BullQueueAdapter } from '../../../../src/jobs/adapter/BullQueueAdapter';
 import type { Job } from '../../../../src/jobs/Job';
-import type { Logger } from '../../../../src/logging/Logger';
-import { getLoggerFor } from '../../../../src/logging/LogUtil';
+import type { BullQueueProcessor } from '../../../../src/jobs/processor/BullQueueProcessor';
 
 jest.mock('bull');
-
-jest.mock('../../../../src/logging/LogUtil', (): any => {
-  const logger: Logger = { info: jest.fn() } as any;
-  return { getLoggerFor: (): Logger => logger };
-});
-const logger: jest.Mocked<Logger> = getLoggerFor('StreamUtil') as any;
 
 Date.now = jest.fn((): number => 1659490735295);
 const dayInMs = 1000 * 60 * 60 * 24;
@@ -23,23 +16,17 @@ describe('A BullQueueAdapter', (): void => {
   let perform: Job['perform'];
   let job: Job;
   let jobs: Record<string, Job>;
-  let process: any;
   let close: any;
   let add: any;
-  let on: any;
   let obliterate: any;
   const redisConfig = { port: 6379, host: '127.0.0.1' };
   let adapter: BullQueueAdapter;
-  let registeredJobs: Record<string, (data: any) => Promise<void>>;
-  let registeredEvents: Record<string, (...args: any[]) => void>;
-  let error: Error | undefined;
-  let stalled: boolean;
+  let queueProcessor: BullQueueProcessor;
 
   beforeEach(async(): Promise<void> => {
     jest.clearAllMocks();
 
     queues = [ queue ];
-    registeredJobs = {};
     perform = jest.fn().mockImplementation(async(): Promise<void> => {
       // Do nothing
     });
@@ -48,282 +35,164 @@ describe('A BullQueueAdapter', (): void => {
 
     close = jest.fn();
     obliterate = jest.fn();
-    process = jest.fn().mockImplementation(
-      (jobName: string, concurrency: number, processFn: (bullJob: any) => Promise<void>): void => {
-        registeredJobs[jobName] = processFn;
-      },
-    );
-
-    error = undefined;
-    stalled = false;
-    add = jest.fn().mockImplementation(
-      async(jobName: string, data: any, options: any): Promise<void> => {
-        if (!options.repeat && !options.delay) {
-          const jobDetails = { name: jobName };
-          registeredEvents.active(jobDetails);
-          if (error) {
-            registeredEvents.failed(jobDetails, error);
-            registeredEvents.error(error);
-          } else {
-            if (stalled) {
-              registeredEvents.stalled(jobDetails);
-            }
-            await registeredJobs[jobName]({ data });
-            registeredEvents.completed(jobDetails);
-          }
-        }
-      },
-    );
-
-    registeredEvents = {};
-    on = jest.fn().mockImplementation((event: string, handler: () => void): void => {
-      registeredEvents[event] = handler;
-    });
+    add = jest.fn();
+    queueProcessor = { processJobsOnQueues: jest.fn() } as any;
 
     (Bull as jest.Mock).mockImplementation(
-      (name: string): Bull.Queue => ({ process, add, on, name, close, obliterate } as any),
+      (name: string): Bull.Queue => ({ process, add, name, close, obliterate } as any),
     );
   });
 
   it('initializes bull queues.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await adapter.performLater('example');
     expect(Bull).toHaveBeenCalledTimes(1);
     expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
     expect(add).toHaveBeenCalledTimes(1);
     expect(add).toHaveBeenCalledWith('example', {}, {});
-    expect(on).toHaveBeenCalledTimes(5);
   });
 
   it('initializes queues with only one total concurrent job.', async(): Promise<void> => {
     jobs = { example: job, example2: job };
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     expect(Bull).toHaveBeenCalledTimes(1);
     expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(2);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
-    expect(process.mock.calls[1][0]).toBe('example2');
-    expect(process.mock.calls[1][1]).toBe(0);
-    expect(on).toHaveBeenCalledTimes(5);
   });
 
   it('closes all queues when finalized.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await adapter.performLater('example');
     expect(Bull).toHaveBeenCalledTimes(1);
     expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
     expect(add).toHaveBeenCalledTimes(1);
     expect(add).toHaveBeenCalledWith('example', {}, {});
-    expect(on).toHaveBeenCalledTimes(5);
     await adapter.finalize();
     expect(close).toHaveBeenCalledTimes(1);
   });
 
   it('errors when attempting to perform a job which has not been defined.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.performLater('sendEmail'))
       .rejects.toThrow('Job \'sendEmail\' is not defined');
   });
 
   it('errors when attempting to perform a job on a queue which has not been defined.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.performLater('example', {}, { queue: 'critical' }))
       .rejects.toThrow('Queue \'critical\' is not defined');
   });
 
   it('adds the job to the queue if the job and queue are defined.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     const data = { alpha: 1 };
     await expect(adapter.performLater('example', data))
       .resolves.toBeUndefined();
     expect(Bull).toHaveBeenCalledTimes(1);
     expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
     expect(add).toHaveBeenCalledTimes(1);
     expect(add).toHaveBeenCalledWith('example', data, {});
-    expect(perform).toHaveBeenCalledTimes(1);
-    expect(perform).toHaveBeenCalledWith(data, adapter);
   });
 
   it('adds the job on a cron schedule.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.performLater('example', {}, { every: '5 4 * * *' }))
       .resolves.toBeUndefined();
     expect(Bull).toHaveBeenCalledTimes(1);
     expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
     expect(add).toHaveBeenCalledTimes(1);
     expect(add).toHaveBeenCalledWith('example', {}, { repeat: { cron: '5 4 * * *' }});
-    expect(perform).toHaveBeenCalledTimes(0);
   });
 
   it('adds the job on a cron schedule with a start time at a specific date.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.performLater('example', {}, { at: tomorrow, every: '5 4 * * *' }))
       .resolves.toBeUndefined();
     expect(Bull).toHaveBeenCalledTimes(1);
     expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
     expect(add).toHaveBeenCalledTimes(1);
     expect(add).toHaveBeenCalledWith('example', {}, { repeat: { cron: '5 4 * * *', startDate: dayInMs }});
-    expect(perform).toHaveBeenCalledTimes(0);
   });
 
   it('adds the job on a cron schedule with a start time after a delay.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.performLater('example', {}, { in: 1000, every: '5 4 * * *' }))
       .resolves.toBeUndefined();
     expect(Bull).toHaveBeenCalledTimes(1);
     expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
     expect(add).toHaveBeenCalledTimes(1);
     expect(add).toHaveBeenCalledWith('example', {}, { repeat: { cron: '5 4 * * *', startDate: 1000 }});
-    expect(perform).toHaveBeenCalledTimes(0);
   });
 
   it('adds the job on a certain millisecond schedule.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.performLater('example', {}, { every: 1000 }))
       .resolves.toBeUndefined();
     expect(Bull).toHaveBeenCalledTimes(1);
     expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
     expect(add).toHaveBeenCalledTimes(1);
     expect(add).toHaveBeenCalledWith('example', {}, { repeat: { every: 1000 }});
-    expect(perform).toHaveBeenCalledTimes(0);
   });
 
   it('adds the job to be run at a specific date.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.performLater('example', {}, { at: tomorrow }))
       .resolves.toBeUndefined();
     expect(Bull).toHaveBeenCalledTimes(1);
     expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
     expect(add).toHaveBeenCalledTimes(1);
     expect(add).toHaveBeenCalledWith('example', {}, { delay: dayInMs });
-    expect(perform).toHaveBeenCalledTimes(0);
   });
 
   it('adds the job to be run after a delay.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.performLater('example', {}, { in: 1000 }))
       .resolves.toBeUndefined();
     expect(Bull).toHaveBeenCalledTimes(1);
     expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
     expect(add).toHaveBeenCalledTimes(1);
     expect(add).toHaveBeenCalledWith('example', {}, { delay: 1000 });
-    expect(perform).toHaveBeenCalledTimes(0);
   });
 
   it('adds the job with a backoff setting.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.performLater('example', {}, { retry: true }))
       .resolves.toBeUndefined();
     expect(Bull).toHaveBeenCalledTimes(1);
     expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
     expect(add).toHaveBeenCalledTimes(1);
     expect(add).toHaveBeenCalledWith('example', {}, { backoff: { type: 'exponential', delay: 2000 }});
-    expect(perform).toHaveBeenCalledTimes(1);
-    expect(perform).toHaveBeenCalledWith({}, adapter);
-  });
-
-  it('logs a message about a job starting then erroring.', async(): Promise<void> => {
-    error = new Error('Job failed');
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
-    await adapter.performLater('example');
-    expect(Bull).toHaveBeenCalledTimes(1);
-    expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
-    expect(add).toHaveBeenCalledTimes(1);
-    expect(add).toHaveBeenCalledWith('example', {}, {});
-    expect(on).toHaveBeenCalledTimes(5);
-    expect(logger.info).toHaveBeenCalledTimes(3);
-    expect(logger.info).toHaveBeenNthCalledWith(1, 'Job example has started on queue default');
-    expect(logger.info.mock.calls[1][0].startsWith('Job example on queue default failed with reason: Job failed'))
-      .toBe(true);
-    expect(logger.info.mock.calls[2][0].startsWith('An error occured in queue default: Job failed'))
-      .toBe(true);
-  });
-
-  it('logs a message about a job starting then completing.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
-    await adapter.performLater('example');
-    expect(Bull).toHaveBeenCalledTimes(1);
-    expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
-    expect(add).toHaveBeenCalledTimes(1);
-    expect(add).toHaveBeenCalledWith('example', {}, {});
-    expect(on).toHaveBeenCalledTimes(5);
-    expect(logger.info).toHaveBeenCalledTimes(2);
-    expect(logger.info).toHaveBeenNthCalledWith(1, 'Job example has started on queue default');
-    expect(logger.info).toHaveBeenNthCalledWith(2, 'Job example successfully completed on queue default');
-  });
-
-  it('logs a message about a job starting then becoming stalled then completing.', async(): Promise<void> => {
-    stalled = true;
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
-    await adapter.performLater('example');
-    expect(Bull).toHaveBeenCalledTimes(1);
-    expect(Bull).toHaveBeenCalledWith('default', { redis: redisConfig });
-    expect(process).toHaveBeenCalledTimes(1);
-    expect(process.mock.calls[0][0]).toBe('example');
-    expect(process.mock.calls[0][1]).toBe(1);
-    expect(add).toHaveBeenCalledTimes(1);
-    expect(add).toHaveBeenCalledWith('example', {}, {});
-    expect(on).toHaveBeenCalledTimes(5);
-    expect(logger.info).toHaveBeenCalledTimes(3);
-    expect(logger.info).toHaveBeenNthCalledWith(1, 'Job example has started on queue default');
-    expect(logger.info).toHaveBeenNthCalledWith(2, 'Job example has been marked as stalled on queue default');
-    expect(logger.info).toHaveBeenNthCalledWith(3, 'Job example successfully completed on queue default');
   });
 
   it('deletes a queue.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.deleteQueue('default')).resolves.toBeUndefined();
     expect(obliterate).toHaveBeenCalledTimes(1);
     expect(obliterate).toHaveBeenCalledWith({ force: true });
   });
 
   it('throws an error when deleting a queue that does not exist.', async(): Promise<void> => {
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.deleteQueue('otherQueue')).rejects.toThrow('No queue named otherQueue found');
     expect(obliterate).toHaveBeenCalledTimes(0);
   });
 
   it('deletes all queues.', async(): Promise<void> => {
     queues = [ queue, 'secondQueue' ];
-    adapter = new BullQueueAdapter({ jobs, queues, redisConfig });
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
     await expect(adapter.deleteAllQueues()).resolves.toBeUndefined();
     expect(obliterate).toHaveBeenCalledTimes(2);
     expect(obliterate).toHaveBeenCalledWith({ force: true });
+  });
+
+  it('sets up jobs to be processed on the queueProcessor.', async(): Promise<void> => {
+    adapter = new BullQueueAdapter({ jobs, queues, redisConfig, queueProcessor });
+    expect(queueProcessor.processJobsOnQueues).toHaveBeenCalledTimes(1);
+    expect(queueProcessor.processJobsOnQueues).toHaveBeenCalledWith(
+      expect.objectContaining({
+        default: expect.any(Object),
+      }),
+      { example: job },
+      adapter,
+    );
   });
 });
